+++ CHANGELOG.md
@@ -13,6 +13,10 @@
   supports source-to-source transformation (and it offers new capabilities to
   AST transform authors that we would like to be available everywhere).
 
+## `@embroider/shared-internals`
+
+- BREAKING: options format of `hbsToJS` changed
+
 # Release 2022-10-06.0
 
 ## `@embroider/core` 1.8.3 -> 1.9.0 minor